+++ fastv/Services/ONNXRuntimeBridge.h
@@ -14,9 +14,5 @@
 #import "onnxruntime_c_api.h"
 #import "KaldiFbankBridge.h"
 
-// 导入 LibEtPan wrapper（用于邮件功能）
-// 使用相对路径，因为这个文件在 fastv/Services/ 目录下
-#import "LibEtPanWrapper.h"
-
 #endif /* ONNXRuntimeBridge_h */